CTCTRADERS-691: Refactored files.
